move constants to properly locations
