+++ src/shots/bunkShoot.ts
@@ -1,9 +1,8 @@
 import type { Bunker } from '@/planet/types'
 import { BunkerKind } from '@/planet/types'
 import type { ShotRec } from './types'
-import { SHOT } from './constants'
+import { SHOT, xbshotstart, ybshotstart } from './constants'
 import { SCRWTH, SOFTBORDER } from '@/screen/constants'
-import { xbshotstart, ybshotstart } from '@/figs/hardcodedSprites'
 import { PLANET } from '@/planet/constants'
 import { aimBunk } from './aimBunk'
 import { aimDir } from './aimDir'
